Fix
사운드 조절 스크립트, 옵션창 왓다갔다
