--- Assets/Scenes/TestScene/SoundSceneTestKyw.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c326e50d0bf4d0e429638b78593b4091
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
