+++ Assets/Scripts/SoundMangerOption.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e12ac4fd5f447ac47b6e2372dbe50cd4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
