+++ Assets/Scripts/SoundMangerOption/GoToStartScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 529588e9dd5210141a3ac93760fe43fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
